rm cloud end to end test
