+++ ambry-vcr/src/test/java/com/github/ambry/vcr/VcrTestUtil.java
@@ -18,6 +18,7 @@ import com.github.ambry.cloud.HelixVcrClusterAgentsFactory;
 import com.github.ambry.cloud.OnlineOfflineHelixVcrStateModelFactory;
 import com.github.ambry.clustermap.ClusterAgentsFactory;
 import com.github.ambry.clustermap.ClusterMap;
+import com.github.ambry.clustermap.DataNodeId;
 import com.github.ambry.clustermap.PartitionId;
 import com.github.ambry.config.CloudConfig;
 import com.github.ambry.config.SSLConfig;
@@ -46,6 +47,9 @@ import org.apache.helix.model.builder.HelixConfigScopeBuilder;
 import org.apache.helix.tools.ClusterSetup;
 import org.apache.helix.zookeeper.api.client.HelixZkClient;
 import org.apache.helix.zookeeper.impl.factory.DedicatedZkClientFactory;
+import org.junit.Test;
+
+import static org.junit.Assume.*;
 
 
 /**
@@ -283,6 +287,40 @@ public class VcrTestUtil {
     }
     vcrServer.shutdown();
     helixControllerManager.syncStop();
+  }
+
+  @Test
+  public void endToEndCloudBackupTest() throws Exception {
+    assumeTrue(testEncryption);
+    sslCluster.startServers();
+    DataNodeId dataNode = sslCluster.getClusterMap().getDataNodeIds().get(0);
+    // Start ZK Server.
+    int zkPort = 31998;
+    String zkConnectString = "localhost:" + zkPort;
+    String vcrClusterName = "vcrTestClusterSSL";
+    TestUtils.ZkInfo zkInfo = new TestUtils.ZkInfo(TestUtils.getTempDir("helixVcr"), "DC1", (byte) 1, zkPort, true);
+    ServerTestUtil.endToEndCloudBackupTest(sslCluster, zkConnectString, vcrClusterName, dataNode, clientSSLConfig2,
+        clientSSLSocketFactory2, notificationSystem, serverSSLProps, false);
+    ServerTestUtil.endToEndCloudBackupTest(sslCluster, zkConnectString, vcrClusterName, dataNode, clientSSLConfig2,
+        clientSSLSocketFactory2, notificationSystem, serverSSLProps, true);
+    zkInfo.shutdown();
+  }
+   
+  @Test
+  public void endToEndCloudBackupTest() throws Exception {
+    assumeTrue(testEncryption);
+    plaintextCluster.startServers();
+    DataNodeId dataNode = plaintextCluster.getClusterMap().getDataNodeIds().get(0);
+    // Start Helix Controller and ZK Server.
+    int zkPort = 31999;
+    String zkConnectString = "localhost:" + zkPort;
+    String vcrClusterName = "vcrTestClusterPlainText";
+    TestUtils.ZkInfo zkInfo = new TestUtils.ZkInfo(TestUtils.getTempDir("helixVcr"), "DC1", (byte) 1, zkPort, true);
+    ServerTestUtil.endToEndCloudBackupTest(plaintextCluster, zkConnectString, vcrClusterName, dataNode, null, null,
+        notificationSystem, null, false);
+    ServerTestUtil.endToEndCloudBackupTest(plaintextCluster, zkConnectString, vcrClusterName, dataNode, null, null,
+        notificationSystem, null, true);
+    zkInfo.shutdown();
   }
    */
 }
